+++ python/PyQt6/core/auto_generated/tiledscene/qgscesiumutils.sip.in
@@ -13,6 +13,7 @@
 
 
 
+
 class QgsCesiumUtils
 {
 %Docstring(signature="appended")
