+++ src/core/tiledscene/qgscesiumutils.h
@@ -22,6 +22,7 @@
 #include <nlohmann/json_fwd.hpp>
 #include <optional>
 
+#include "qgis.h"
 #include "qgis_core.h"
 #include "qgis_sip.h"
 #include "qgsbox3d.h"
@@ -40,6 +41,14 @@ class QgsSphere;
 class QgsOrientedBox3D;
 class QgsMatrix4x4;
 class QgsTiledSceneBoundingVolume;
+
+#ifndef SIP_RUN
+namespace tinygltf
+{
+  class Model;
+}
+#include "qgsgltfutils.h"
+#endif
 
 /**
  * \brief Contains utilities for working with Cesium data.
@@ -121,11 +130,14 @@ class CORE_EXPORT QgsCesiumUtils
 
 #ifndef SIP_RUN
     /**
-     * \brief Raw per-instance data parsed from an i3dm feature table.
+     * \brief Raw per-instance data parsed from an i3dm feature table of a single tile
      *
      * This struct transports i3dm instance data from the binary parser through
-     * TileContents to resolveInstancing(). EXT_mesh_gpu_instancing does not use
-     * this struct - it is parsed directly from glTF nodes inside resolveInstancing().
+     * TileContents to resolveInstancing() - to handle 3D Tiles 1.0 spec.
+     *
+     * In 3D Tiles 1.1 spec, instanced meshes are handled by EXT_mesh_gpu_instancing
+     * glTF extension, and this struct is not used at all, because instancing is parsed
+     * directly from glTF nodes inside resolveInstancing().
      *
      * All positions are in i3dm tile space (Z-up), relative to RTC_CENTER.
      *
@@ -133,7 +145,7 @@ class CORE_EXPORT QgsCesiumUtils
      *
      * \since QGIS 4.2
      */
-    struct QgsGltfInstancingData
+    struct TileI3dmData
     {
         //! Number of instances
         int instanceCount = 0;
@@ -161,7 +173,7 @@ class CORE_EXPORT QgsCesiumUtils
 
 #ifndef SIP_RUN
         //! Optional instancing data, populated for i3dm tiles
-        std::optional<QgsGltfInstancingData> instancing;
+        std::optional<TileI3dmData> instancing;
 #endif
     };
 
@@ -204,21 +216,27 @@ class CORE_EXPORT QgsCesiumUtils
 #ifndef SIP_RUN
 
     /**
-     * Computes EAST_NORTH_UP orientation quaternions for each instance position.
+     * Resolves instancing from either i3dm data or EXT_mesh_gpu_instancing.
      *
-     * Positions are in tile-local Z-up space, relative to \a rtcCenter.
-     * The \a tileTransform converts from tile-local space to ECEF (EPSG:4978),
-     * so that the ENU frame can be computed at the correct geographic location.
+     * Walks the glTF node tree, decomposes multi-node models into independent
+     * mesh primitives, and computes per-instance TRS in a common "tile space".
      *
-     * \param positions per-instance positions in tile-local space
+     * Returns an empty vector if no instancing data is found, meaning the model
+     * should be rendered through the existing non-instanced code path.
+     *
+     * \param model the loaded tinygltf model
+     * \param tileInstancing optional i3dm instance data (nullopt for EXT/b3dm)
+     * \param gltfUpAxis the up axis used in the glTF model
+     * \param tileTransform tile transform from tileset.json (tile-local → ECEF), needed for deferred EAST_NORTH_UP computation
      * \param rtcCenter RTC_CENTER offset in tile-local space
-     * \param tileTransform tile transform from tileset.json (tile-local → ECEF)
-     * \param rotations output quaternions (one per instance)
+     * \returns flat list of instanced primitives with tile-space transforms
      * \since QGIS 4.2
      */
-    static void computeEastNorthUpQuaternions( const QVector<QVector3D> &positions, const QgsVector3D &rtcCenter, const QgsMatrix4x4 &tileTransform, QVector<QQuaternion> &rotations );
+    static QVector<QgsGltfUtils::InstancedPrimitive> resolveInstancing(
+      const tinygltf::Model &model, const std::optional<TileI3dmData> &tileInstancing, Qgis::Axis gltfUpAxis, const QgsMatrix4x4 &tileTransform, const QgsVector3D &rtcCenter
+    );
 
-#endif
+#endif // SIP_RUN
 };
 
 #endif // QGSCESIUMUTILS_H
